# internal/web/go/model_company.go
/*
 * Swagger Petstore - OpenAPI 3.0
 *
 * Documentation of a simple state machine that handles applications work state
 *
 * API version: 0.0.1
 * Contact: projects@mkdirdev.co.uk
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type Company struct {
	Id int64 `json:"id" bson:"_id"`

	Name string `json:"name" bson:"name"`

	Users []User `json:"users" bson:"users"`
}