# internal/web/go/model_task.go
/*
 * Swagger Petstore - OpenAPI 3.0
 *
 * Documentation of a simple state machine that handles applications work state
 *
 * API version: 0.0.1
 * Contact: projects@mkdirdev.co.uk
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type Task struct {
	Id int64 `json:"id" bson:"id"`

	Name string `json:"name" bson:"name"`

	AssignedUsers []User `json:"assigned_users" bson:"assigned_users"`

	SubTasks []Task `json:"sub_tasks" bson:"sub_tasks"`

	Comments []string `json:"comments" bson:"comments"`

	Completed bool `json:"completed" bson:"completed"`
}