# internal/web/go/model_update_status.go
/*
 * Swagger Petstore - OpenAPI 3.0
 *
 * Documentation of a simple state machine that handles applications work state
 *
 * API version: 0.0.1
 * Contact: projects@mkdirdev.co.uk
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type update_status string

// List of update_status
const (
	UPDATE   update_status = "update"
	COMPLETE update_status = "complete"
	CANCEL   update_status = "cancel"
)