# internal/web/go/model_user.go
/*
 * Swagger Petstore - OpenAPI 3.0
 *
 * Documentation of a simple state machine that handles applications work state
 *
 * API version: 0.0.1
 * Contact: projects@mkdirdev.co.uk
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type User struct {
	Id int64 `json:"id" bson:"Id"`

	UserName string `json:"user_name" bson:"user_name"`

	CompanyId string `json:"company_Id" bson:"companyId"`
}